Add documentation, better formatting, and move private change helper functions to lib/changes.py
