+++ ckan/templates/package/changes.html
@@ -18,18 +18,19 @@
       {% block package_changes_header %}
         <h1 class="page-heading">{{ _('Changes') }}</h1>
       {% endblock %}
+
       {% set select_list1 = h.activity_list_select(pkg_activity_list, activity_diffs[-1].activities[0].id) %}
       {% set select_list2 = h.activity_list_select(pkg_activity_list, activity_diffs[0].activities[1].id) %}
       <form id="range_form" action="{{ h.url_for('dataset.change_range') }}" >
         <input type="hidden" name="current_old_id" value="{{ activity_diffs[-1].activities[0].id }}">
         <input type="hidden" name="current_new_id" value="{{ activity_diffs[0].activities[1].id }}">
       View changes from
-        <select class="form-control" style="width:215px;display:inline;" form="range_form" name="oldest_id">
+        <select class="form-control" style="width:235px;display:inline;" form="range_form" name="oldest_id">
           <pre>
             {{ select_list1[1:]|safe }}
           </pre>
         </select> to
-        <select class="form-control" style="width:215px;display:inline;" form="range_form" name="newest_id">
+        <select class="form-control" style="width:235px;display:inline;" form="range_form" name="newest_id">
           <pre>
             {{ select_list2|safe }}
           </pre>
@@ -39,10 +40,12 @@
 
         <br>
 
+      {# iterate through the list of activity diffs #}
+      <hr>
       {% for i in range(activity_diffs|length) %}
         {% snippet "package/snippets/change_item.html", activity_diff=activity_diffs[i], pkg_dict=pkg_dict %}
         <br style="line-height:2;">
-        <br>
+        <hr>
       {% endfor %}
 
     </div>
